+++ src/middleware.ts
@@ -57,8 +57,8 @@ export const config = {
      * - _next/static (static files)
      * - _next/image (image optimization files)
      * - favicon.ico (favicon file)
-     * - *.svg / *.png / *.jpg / *.jpeg / *.gif (static assets)
+     * - *.svg / *.png / *.jpg / *.jpeg / *.gif / *.webp / *.lottie (static assets)
      */
-    '/((?!api|_next/static|_next/image|favicon.ico|.*\\.svg$|.*\\.png$|.*\\.jpg$|.*\\.jpeg$|.*\\.gif$).*)',
+    '/((?!api|_next/static|_next/image|favicon.ico|.*\\.svg$|.*\\.png$|.*\\.jpg$|.*\\.jpeg$|.*\\.gif$|.*\\.webp$|.*\\.lottie$).*)',
   ],
 }